+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * altered tokio dep, full is now optional (prevents valgrind errors)
 * altered utils dep to 0.1
 * added debug log for when eval/evalmodule fails
+* scriptpreproc returns Result instead of just script
 
 # 0.4.0